Allow customisation of input parameters
This allows to change the way we detect a button is pressed or released.
We were just hardcoding HIGH which is generally fine for push buttons,
but other buttons such as rotary encoders can read a press as LOW.

Additionally allows the button pin to be set as a pullup.